aisi_deployment_api/requirements.txt
anyio==3.3.3
asgiref==3.4.1
azure-common==1.1.28
azure-core==1.24.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0
azure-nspkg==3.0.2
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-file-share==12.8.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
contextlib2==21.6.0
contextvars==2.4
cryptography==35.0.0
dataclasses==0.8
fastapi==0.70.0
h11==0.12.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.1
isodate==0.6.0
msal==1.15.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
portalocker==1.7.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.26.0
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.6.0
